% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dictSP}
\alias{dictSP}
\alias{od_sao_paulo_1977_not_harmonized_dictionary_en}
\alias{od_sao_paulo_1977_not_harmonized_dictionary_es}
\alias{od_sao_paulo_1977_not_harmonized_dictionary_pt}
\alias{od_sao_paulo_1987_not_harmonized_dictionary_en}
\alias{od_sao_paulo_1987_not_harmonized_dictionary_es}
\alias{od_sao_paulo_1987_not_harmonized_dictionary_pt}
\alias{od_sao_paulo_1997_not_harmonized_dictionary_en}
\alias{od_sao_paulo_1997_not_harmonized_dictionary_es}
\alias{od_sao_paulo_1997_not_harmonized_dictionary_pt}
\alias{od_sao_paulo_2007_not_harmonized_dictionary_en}
\alias{od_sao_paulo_2007_not_harmonized_dictionary_es}
\alias{od_sao_paulo_2007_not_harmonized_dictionary_pt}
\alias{od_sao_paulo_2017_not_harmonized_dictionary_en}
\alias{od_sao_paulo_2017_not_harmonized_dictionary_es}
\alias{od_sao_paulo_2017_not_harmonized_dictionary_pt}
\title{Sao Paulo OD Survey Dictionary}
\format{
A data frame with 4 columns:
\describe{
\item{variable_name}{Name of the variable}
\item{description}{Description of the variable}
\item{categories}{Examples of the categories in the variable}
\item{class}{Class of the variable}
}

An object of class \code{data.table} (inherits from \code{data.frame}) with 76 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 76 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 93 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 93 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 93 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 110 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 76 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 110 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 76 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 76 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 76 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 76 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 76 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 76 rows and 4 columns.
}
\source{
\url{https://transparencia.metrosp.com.br/dataset/pesquisa-origem-e-destino}
}
\usage{
od_sao_paulo_1977_not_harmonized_dictionary_en

od_sao_paulo_1977_not_harmonized_dictionary_es

od_sao_paulo_1977_not_harmonized_dictionary_pt

od_sao_paulo_1987_not_harmonized_dictionary_en

od_sao_paulo_1987_not_harmonized_dictionary_es

od_sao_paulo_1987_not_harmonized_dictionary_pt

od_sao_paulo_1997_not_harmonized_dictionary_en

od_sao_paulo_1997_not_harmonized_dictionary_es

od_sao_paulo_1997_not_harmonized_dictionary_pt

od_sao_paulo_2007_not_harmonized_dictionary_en

od_sao_paulo_2007_not_harmonized_dictionary_es

od_sao_paulo_2007_not_harmonized_dictionary_pt

od_sao_paulo_2017_not_harmonized_dictionary_en

od_sao_paulo_2017_not_harmonized_dictionary_es

od_sao_paulo_2017_not_harmonized_dictionary_pt
}
\description{
These datasets contain the dictionary for the OD surveys in Sao Paulo.
Each row describes one column of the survey data frame.
}
\keyword{Dictionary}
\keyword{datasets}
