% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{Metadata for the package}
\format{
\subsection{\code{metadata}}{

A data frame with 4 columns:
\describe{
\item{city}{City name}
\item{year}{Year of the survey}
\item{harmonized}{A logical value showing whether the dataset was harmonized}
\item{language}{Language of the dictionary}
}
}
}
\usage{
metadata
}
\description{
This dataset has the list of OD surveys available
in the package.
}
\keyword{datasets}
