% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_piggyback.R
\name{download_piggyback}
\alias{download_piggyback}
\title{Download file from odbr using piggyback}
\usage{
download_piggyback(filename_to_download)
}
\arguments{
\item{filename_to_download}{String. The name of the file to be downloaded.}
}
\value{
A string with the path to the file saved in a tempdir
}
\description{
Download file from odbr using piggyback
}
\keyword{internal}
