% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_map.R
\name{read_map}
\alias{read_map}
\title{Download spatial data from OD Surveys databases}
\usage{
read_map(
  city = "São Paulo",
  year = 2017,
  harmonize = FALSE,
  geometry = "zone",
  force_download = FALSE
)
}
\arguments{
\item{city}{Character. City of reference. Defaults to "São Paulo".}

\item{year}{Numeric. Year of reference in the format \code{yyyy}. Defaults to
\code{1977}.}

\item{harmonize}{Logical. When \code{FALSE} (Default), the function returns the raw
data. If \code{TRUE}, the function returns harmonized data to the same
city, across all the years.}

\item{geometry}{Character. The type of spatial data to be opened. Options
include \code{c("zone","district", "municipality")}.}

\item{force_download}{Boolean. If we should download the data file again even
if it was already downloaded. Defaults to FALSE.}
}
\value{
An \verb{"sf" "data.frame"} object
}
\description{
The "read_map" function requires as parameter city, year, geometry (zones or
municipalities) - the default is zone, and whether you want the harmonized
database (variable names and reference/geodesic projection over the years,
for the same city) or not - the default is the raw base.
}
\examples{
\dontshow{if (identical(tolower(Sys.getenv("NOT_CRAN")), "true")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(odbr)

# return zone data from OD Surveys database as sf object at a given city and year
df <- read_map(
  city = "Sao Paulo",
  year = 2017,
  harmonize = FALSE,
  geometry = "zone"
)

#' # return district data from OD Surveys database as sf object at a given city and year
df <- read_map(
  city = "Sao Paulo",
  year = 2017,
  harmonize = FALSE,
  geometry = "district"
)

# return municipality data from OD Surveys database as sf object at a given city and year
df <- read_map(
  city = "Sao Paulo",
  year = 2017,
  harmonize = FALSE,
  geometry = "municipality"
)
\dontshow{\}) # examplesIf}
}
\concept{Geometry}
