% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_map.R
\name{read_map}
\alias{read_map}
\title{Download spatial data from OD Surveys databases}
\usage{
read_map(city = "São Paulo", year = 1977, harmonize = FALSE, geometry = "zone")
}
\arguments{
\item{city}{Character. City of reference. Defaults to "São Paulo".}

\item{year}{Numeric. Year of reference in the format \code{yyyy}. Defaults to
\code{1977}.}

\item{harmonize}{Logical. When \code{FALSE} (Default), the function returns the raw
data. If \code{TRUE}, the function returns harmonized data to the same
city, across all the years.}

\item{geometry}{Character. The type of spatial data to be opened. Options
include \code{c("zone","district", "municipality")}.}
}
\value{
An \verb{"sf" "data.frame"} object
}
\description{
The "read_map" function requires as parameter city, year, geometry (zones or
municipalities) - the default is zone, and whether you want the harmonized
database (variable names and reference/geodesic projection over the years,
for the same city) or not - the default is the raw base.
}
\examples{
library(odbr)

# return zone data from OD Surveys database as data.frame at a given city and year
df <- read_map(
  city = "São Paulo",
  year = 1977,
  harmonize = FALSE,
  geometry = "zone"
)

#' # return district data from OD Surveys database as data.frame at a given city and year
df <- read_map(
  city = "São Paulo",
  year = 1977,
  harmonize = FALSE,
  geometry = "district"
)

# return municipality data from OD Surveys database as data.frame at a given city and year
df <- read_map(
  city = "São Paulo",
  year = 1977,
  harmonize = FALSE,
  geometry = "municipality"
)

}
\concept{Geometry}
